fix: get_time_ms()
キャストを明示

+++ philo/utils.c
@@ -44,7 +44,7 @@ int64_t	get_time_ms(void)
 	struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
-	return (tv.tv_sec * 1000 + tv.tv_usec / 1000);
+	return ((int64_t)tv.tv_sec * 1000 + tv.tv_usec / 1000);
 }
 
 void	ft_usleep(int64_t ms)
